+++ lib/src/common/binance_abstract.dart
@@ -51,7 +51,10 @@ abstract class BinanceAbstract {
     Map<String, String>? params,
   }) async {
     params ??= {};
-    if (timestampRequired) params['timestamp'] = (DateTime.now().millisecondsSinceEpoch - timestampDifference).toString();
+    if (timestampRequired)
+      params['timestamp'] =
+          (DateTime.now().millisecondsSinceEpoch - timestampDifference)
+              .toString();
 
     if (signatureRequired) {
       if (_apiSecret == null) {
@@ -108,7 +111,8 @@ abstract class BinanceAbstract {
 
     if (result is Map) {
       if (result.containsKey("code") && result['code'] != 200) {
-        return Left("Binance API returned error ${result["code"]} : ${result["msg"]}");
+        return Left(
+            "Binance API returned error ${result["code"]} : ${result["msg"]}");
       }
     }