+++ lib/src/spot/classes.dart
@@ -27,7 +27,10 @@ class Symbol {
         quoteAsset = m['quoteAsset'],
         quotePrecision = m['quotePrecision'],
         quoteAssetPrecision = m['quoteAssetPrecision'],
-        orderTypes = (m['orderTypes'] as List<dynamic>).cast<String>().map((String e) => e.toOrderTypeEnum()).toList(),
+        orderTypes = (m['orderTypes'] as List<dynamic>)
+            .cast<String>()
+            .map((String e) => e.toOrderTypeEnum())
+            .toList(),
         icebergAllowed = m['icebergAllowed'],
         ocoAllowed = m['ocoAllowed'],
         isSpotTradingAllowed = m['isSpotTradingAllowed'],
@@ -48,7 +51,9 @@ class ExchangeInfo {
         serverTime = m['serverTime'],
         rateLimits = m['rateLimits'],
         exchangeFilters = m['exchangeFilters'],
-        symbols = (m['symbols'] as List<dynamic>).map((e) => Symbol.fromMap(e)).toList();
+        symbols = (m['symbols'] as List<dynamic>)
+            .map((e) => Symbol.fromMap(e))
+            .toList();
 }
 
 class OrderBook {
@@ -62,8 +67,12 @@ class OrderBook {
       : lastUpdateId = m['lastUpdateId'],
         messageOutputTime = m['E'],
         transactionTime = m['T'],
-        bids = (m['bids'] as List<dynamic>).map((e) => DepthOrder.fromList(e)).toList(),
-        asks = (m['asks'] as List<dynamic>).map((e) => DepthOrder.fromList(e)).toList();
+        bids = (m['bids'] as List<dynamic>)
+            .map((e) => DepthOrder.fromList(e))
+            .toList(),
+        asks = (m['asks'] as List<dynamic>)
+            .map((e) => DepthOrder.fromList(e))
+            .toList();
 }
 
 class Trade {
@@ -317,12 +326,19 @@ class OcoOrder {
       : orderListId = m['orderListId'],
         contingencyType = m['contingencyType'],
         listStatusType = (m['listStatusType'] as String).toOcoStatusEnum(),
-        listOrderStatus = (m['listOrderStatus'] as String).toOcoOrderStatusEnum(),
+        listOrderStatus =
+            (m['listOrderStatus'] as String).toOcoOrderStatusEnum(),
         listClientOrderId = m['listClientOrderId'],
         transactionTime = m['transactionTime'],
         symbol = m['symbol'],
-        orders = (m['orders'] as List<dynamic>).map((e) => OrderInfo.fromMap(e)).toList(),
-        orderReports = m.containsKey('orderReports') ? (m['orderReports'] as List<dynamic>).map((e) => Order.fromMap(e)).toList() : [];
+        orders = (m['orders'] as List<dynamic>)
+            .map((e) => OrderInfo.fromMap(e))
+            .toList(),
+        orderReports = m.containsKey('orderReports')
+            ? (m['orderReports'] as List<dynamic>)
+                .map((e) => Order.fromMap(e))
+                .toList()
+            : [];
 }
 
 class Balance {
@@ -359,7 +375,9 @@ class AccountInfo {
         canDeposit = m['canDeposit'],
         updateTime = m['updateTime'],
         accountType = m['accountType'],
-        balances = (m['balances'] as List<dynamic>).map((e) => Balance.fromMap(e)).toList(),
+        balances = (m['balances'] as List<dynamic>)
+            .map((e) => Balance.fromMap(e))
+            .toList(),
         permissions = m['permissions'].cast<String>();
 }
 
@@ -427,7 +445,9 @@ class FiatTransferHistory {
   FiatTransferHistory.fromMap(Map m)
       : code = double.parse(m['code']),
         message = m['message'],
-        data = (m['data'] as List<dynamic>).map((e) => FiatTransferData.fromMap(e)).toList(),
+        data = (m['data'] as List<dynamic>)
+            .map((e) => FiatTransferData.fromMap(e))
+            .toList(),
         total = m['total'],
         success = m['success'];
 }
@@ -467,7 +487,9 @@ class FiatPaymentHistory {
   FiatPaymentHistory.fromMap(Map m)
       : code = double.parse(m['code']),
         message = m['message'],
-        data = (m['data'] as List<dynamic>).map((e) => FiatPaymentData.fromMap(e)).toList(),
+        data = (m['data'] as List<dynamic>)
+            .map((e) => FiatPaymentData.fromMap(e))
+            .toList(),
         total = m['total'],
         success = m['success'];
 }
@@ -486,7 +508,10 @@ class SubaccountInfo {
 class SubaccountList {
   List<SubaccountInfo> subAccounts;
 
-  SubaccountList.fromMap(Map m) : subAccounts = (m['subAccounts'] as List<dynamic>).map((e) => SubaccountInfo.fromMap(e)).toList();
+  SubaccountList.fromMap(Map m)
+      : subAccounts = (m['subAccounts'] as List<dynamic>)
+            .map((e) => SubaccountInfo.fromMap(e))
+            .toList();
 }
 
 class SubaccountTransfer {
@@ -516,7 +541,9 @@ class SubaccountFuturesTransferHistory {
   SubaccountFuturesTransferHistory.fromMap(Map m)
       : success = m['success'],
         futuresType = m['futuresType'],
-        transfers = (m['transfers'] as List<dynamic>).map((e) => SubaccountTransfer.fromMap(e)).toList();
+        transfers = (m['transfers'] as List<dynamic>)
+            .map((e) => SubaccountTransfer.fromMap(e))
+            .toList();
 }
 
 class SubaccountFuturesTransfer {
@@ -531,7 +558,10 @@ class SubaccountFuturesTransfer {
 class SubaccountBalances {
   List<Balance> balances;
 
-  SubaccountBalances.fromMap(Map m) : balances = (m['balances'] as List<dynamic>).map((e) => Balance.fromMap(e)).toList();
+  SubaccountBalances.fromMap(Map m)
+      : balances = (m['balances'] as List<dynamic>)
+            .map((e) => Balance.fromMap(e))
+            .toList();
 }
 
 class SubaccountBtcValued {
@@ -552,7 +582,9 @@ class SubaccountSpotAssetsSummary {
       : totalCount = m['totalCount'],
         masterAccountTotalAsset = double.parse(m['masterAccountTotalAsset']),
         spotSubUserAssetBtcVoList =
-            (m['spotSubUserAssetBtcVoList'] as List<dynamic>).map((e) => SubaccountBtcValued.fromMap(e)).toList();
+            (m['spotSubUserAssetBtcVoList'] as List<dynamic>)
+                .map((e) => SubaccountBtcValued.fromMap(e))
+                .toList();
 }
 
 class DepositAddress {
@@ -657,8 +689,12 @@ class SubaccountMarginDetail {
         totalAssetOfBtc = double.parse(m['totalAssetOfBtc']),
         totalLiabilityOfBtc = double.parse(m['totalLiabilityOfBtc']),
         totalNetAssetOfBtc = double.parse(m['totalNetAssetOfBtc']),
-        marginTradeCoeffVo = (m['marginTradeCoeffVo'] as List<dynamic>).map((e) => SubaccountMarginCoeff.fromMap(e)).toList(),
-        marginUserAssetVoList = (m['marginUserAssetVoList'] as List<dynamic>).map((e) => MarginBalance.fromMap(e)).toList();
+        marginTradeCoeffVo = (m['marginTradeCoeffVo'] as List<dynamic>)
+            .map((e) => SubaccountMarginCoeff.fromMap(e))
+            .toList(),
+        marginUserAssetVoList = (m['marginUserAssetVoList'] as List<dynamic>)
+            .map((e) => MarginBalance.fromMap(e))
+            .toList();
 }
 
 class SubaccountMarginSummary {
@@ -684,7 +720,9 @@ class AllSubaccountMarginSummary {
       : totalAssetOfBtc = double.parse(m['totalAssetOfBtc']),
         totalLiabilityOfBtc = double.parse(m['totalLiabilityOfBtc']),
         totalNetAssetOfBtc = double.parse(m['totalNetAssetOfBtc']),
-        subAccountList = (m['subAccountList'] as List<dynamic>).map((e) => SubaccountMarginSummary.fromMap(e)).toList();
+        subAccountList = (m['subAccountList'] as List<dynamic>)
+            .map((e) => SubaccountMarginSummary.fromMap(e))
+            .toList();
 }
 
 class SubaccountFuturesAsset {
@@ -731,7 +769,9 @@ class SubaccountFuturesDetail {
   SubaccountFuturesDetail.fromMap(Map m)
       : email = m['email'],
         asset = m['asset'],
-        assets = (m['assets'] as List<dynamic>).map((e) => SubaccountFuturesAsset.fromMap(e)).toList(),
+        assets = (m['assets'] as List<dynamic>)
+            .map((e) => SubaccountFuturesAsset.fromMap(e))
+            .toList(),
         canDeposit = m['canDeposit'],
         canTrade = m['canTrade'],
         canWithdraw = m['canWithdraw'],
@@ -740,8 +780,10 @@ class SubaccountFuturesDetail {
         totalInitialMargin = double.parse(m['totalInitialMargin']),
         totalMaintenanceMargin = double.parse(m['totalMaintenanceMargin']),
         totalMarginBalance = double.parse(m['totalMarginBalance']),
-        totalOpenOrderInitialMargin = double.parse(m['totalOpenOrderInitialMargin']),
-        totalPositionInitialMargin = double.parse(m['totalPositionInitialMargin']),
+        totalOpenOrderInitialMargin =
+            double.parse(m['totalOpenOrderInitialMargin']),
+        totalPositionInitialMargin =
+            double.parse(m['totalPositionInitialMargin']),
         totalUnrealizedProfit = double.parse(m['totalUnrealizedProfit']),
         totalWalletBalance = double.parse(m['totalWalletBalance']),
         updateTime = m['updateTime'];
@@ -763,8 +805,10 @@ class SubaccountFuturesSummary {
         totalInitialMargin = double.parse(m['totalInitialMargin']),
         totalMaintenanceMargin = double.parse(m['totalMaintenanceMargin']),
         totalMarginBalance = double.parse(m['totalMarginBalance']),
-        totalOpenOrderInitialMargin = double.parse(m['totalOpenOrderInitialMargin']),
-        totalPositionInitialMargin = double.parse(m['totalPositionInitialMargin']),
+        totalOpenOrderInitialMargin =
+            double.parse(m['totalOpenOrderInitialMargin']),
+        totalPositionInitialMargin =
+            double.parse(m['totalPositionInitialMargin']),
         totalUnrealizedProfit = double.parse(m['totalUnrealizedProfit']),
         totalWalletBalance = double.parse(m['totalWalletBalance']),
         asset = m['asset'];
@@ -785,12 +829,16 @@ class AllSubaccountFuturesSummary {
       : totalInitialMargin = double.parse(m['totalInitialMargin']),
         totalMaintenanceMargin = double.parse(m['totalMaintenanceMargin']),
         totalMarginBalance = double.parse(m['totalMarginBalance']),
-        totalOpenOrderInitialMargin = double.parse(m['totalOpenOrderInitialMargin']),
-        totalPositionInitialMargin = double.parse(m['totalPositionInitialMargin']),
+        totalOpenOrderInitialMargin =
+            double.parse(m['totalOpenOrderInitialMargin']),
+        totalPositionInitialMargin =
+            double.parse(m['totalPositionInitialMargin']),
         totalUnrealizedProfit = double.parse(m['totalUnrealizedProfit']),
         totalWalletBalance = double.parse(m['totalWalletBalance']),
         asset = m['asset'],
-        subAccountList = (m['subAccountList'] as List<dynamic>).map((e) => SubaccountFuturesSummary.fromMap(e)).toList();
+        subAccountList = (m['subAccountList'] as List<dynamic>)
+            .map((e) => SubaccountFuturesSummary.fromMap(e))
+            .toList();
 }
 
 class SubaccountFuturesPositionRisk {
@@ -957,7 +1005,9 @@ class CoinData {
         isLegalMoney = m['isLegalMoney'],
         locked = double.parse(m['locked']),
         name = m['name'],
-        networkList = (m['networkList'] as List<dynamic>).map((e) => Network.fromMap(e)).toList(),
+        networkList = (m['networkList'] as List<dynamic>)
+            .map((e) => Network.fromMap(e))
+            .toList(),
         storage = double.parse(m['storage']),
         trading = m['trading'],
         withdrawAllEnable = m['withdrawAllEnable'],
@@ -969,7 +1019,9 @@ class SnapshotData {
   double totalAssetOfBtc;
 
   SnapshotData.fromMap(Map m)
-      : balances = (m['balances'] as List<dynamic>).map((e) => Balance.fromMap(e)).toList(),
+      : balances = (m['balances'] as List<dynamic>)
+            .map((e) => Balance.fromMap(e))
+            .toList(),
         totalAssetOfBtc = double.parse(m['totalAssetOfBtc']);
 }
 
@@ -992,7 +1044,9 @@ class Snapshots {
   Snapshots.fromMap(Map m)
       : code = m['code'],
         msg = m['msg'],
-        snapshotVos = (m['snapshotVos'] as List<dynamic>).map((e) => SnapshotVos.fromMap(e)).toList();
+        snapshotVos = (m['snapshotVos'] as List<dynamic>)
+            .map((e) => SnapshotVos.fromMap(e))
+            .toList();
 }
 
 class SnapshotMarginData {
@@ -1007,7 +1061,9 @@ class SnapshotMarginData {
         totalAssetOfBtc = double.parse(m['totalAssetOfBtc']),
         totalLiabilityOfBtc = double.parse(m['totalLiabilityOfBtc']),
         totalNetAssetOfBtc = double.parse(m['totalNetAssetOfBtc']),
-        userAssets = (m['userAssets'] as List<dynamic>).map((e) => MarginBalance.fromMap(e)).toList();
+        userAssets = (m['userAssets'] as List<dynamic>)
+            .map((e) => MarginBalance.fromMap(e))
+            .toList();
 }
 
 class SnapshotVosMargin {
@@ -1029,7 +1085,9 @@ class SnapshotsMargin {
   SnapshotsMargin.fromMap(Map m)
       : code = m['code'],
         msg = m['msg'],
-        snapshotVos = (m['snapshotVos'] as List<dynamic>).map((e) => SnapshotVosMargin.fromMap(e)).toList();
+        snapshotVos = (m['snapshotVos'] as List<dynamic>)
+            .map((e) => SnapshotVosMargin.fromMap(e))
+            .toList();
 }
 
 class FuturesBalance {
@@ -1063,8 +1121,12 @@ class SnapshotFuturesData {
   List<FuturesPosition> position;
 
   SnapshotFuturesData.fromMap(Map m)
-      : assets = (m['assets'] as List<dynamic>).map((e) => FuturesBalance.fromMap(e)).toList(),
-        position = (m['position'] as List<dynamic>).map((e) => FuturesPosition.fromMap(e)).toList();
+      : assets = (m['assets'] as List<dynamic>)
+            .map((e) => FuturesBalance.fromMap(e))
+            .toList(),
+        position = (m['position'] as List<dynamic>)
+            .map((e) => FuturesPosition.fromMap(e))
+            .toList();
 }
 
 class SnapshotVosFutures {
@@ -1086,7 +1148,9 @@ class SnapshotsFutures {
   SnapshotsFutures.fromMap(Map m)
       : code = m['code'],
         msg = m['msg'],
-        snapshotVos = (m['snapshotVos'] as List<dynamic>).map((e) => SnapshotVosFutures.fromMap(e)).toList();
+        snapshotVos = (m['snapshotVos'] as List<dynamic>)
+            .map((e) => SnapshotVosFutures.fromMap(e))
+            .toList();
 }
 
 class Withdraw {
@@ -1155,7 +1219,10 @@ class AccountApiTradingData {
       : isLocked = m['isLocked'],
         plannedRecoverTime = m['plannedRecoverTime'],
         triggerCondition = m['triggerCondition'],
-        indicators = (m['indicators'] as Map<String, dynamic>).entries.map((e) => ApiTradingIndicator.fromEntry(e)).toList(),
+        indicators = (m['indicators'] as Map<String, dynamic>)
+            .entries
+            .map((e) => ApiTradingIndicator.fromEntry(e))
+            .toList(),
         updateTime = m['updateTime'];
 }
 
@@ -1188,7 +1255,10 @@ class Dribblet {
         totalTransferedAmount = double.parse(m['totalTransferedAmount']),
         totalServiceChargeAmount = double.parse(m['totalServiceChargeAmount']),
         transId = m['transId'],
-        userAssetDribbletDetails = (m['userAssetDribbletDetails'] as List<dynamic>).map((e) => DribbletDetail.fromMap(e)).toList();
+        userAssetDribbletDetails =
+            (m['userAssetDribbletDetails'] as List<dynamic>)
+                .map((e) => DribbletDetail.fromMap(e))
+                .toList();
 }
 
 class DustLog {
@@ -1197,7 +1267,9 @@ class DustLog {
 
   DustLog.fromMap(Map m)
       : total = m['total'],
-        userAssetDribblets = (m['userAssetDribblets'] as List<dynamic>).map((e) => Dribblet.fromMap(e)).toList();
+        userAssetDribblets = (m['userAssetDribblets'] as List<dynamic>)
+            .map((e) => Dribblet.fromMap(e))
+            .toList();
 }
 
 class DustConversion {
@@ -1208,7 +1280,9 @@ class DustConversion {
   DustConversion.fromMap(Map m)
       : totalServiceCharge = double.parse(m['totalServiceCharge']),
         totalTransfered = double.parse(m['totalTransfered']),
-        transferResult = (m['transferResult'] as List<dynamic>).map((e) => DribbletDetail.fromMap(e)).toList();
+        transferResult = (m['transferResult'] as List<dynamic>)
+            .map((e) => DribbletDetail.fromMap(e))
+            .toList();
 }
 
 class Dividend {
@@ -1231,7 +1305,9 @@ class DividendRecord {
   int total;
 
   DividendRecord.fromMap(Map m)
-      : rows = (m['rows'] as List<dynamic>).map((e) => Dividend.fromMap(e)).toList(),
+      : rows = (m['rows'] as List<dynamic>)
+            .map((e) => Dividend.fromMap(e))
+            .toList(),
         total = m['total'];
 }
 
@@ -1286,7 +1362,9 @@ class UniversalTransferHistory {
 
   UniversalTransferHistory.fromMap(Map m)
       : total = m['total'],
-        rows = (m['rows'] as List<dynamic>).map((e) => UniversalTransfer.fromMap(e)).toList();
+        rows = (m['rows'] as List<dynamic>)
+            .map((e) => UniversalTransfer.fromMap(e))
+            .toList();
 }
 
 class FundingAsset {
@@ -1343,7 +1421,8 @@ class WsAccountUpdate {
       : eventType = m['e'],
         eventTime = m['E'],
         lastUpdateTime = m['u'],
-        balances = (m['B'] as List<dynamic>).map((e) => Balance.fromMap(e)).toList();
+        balances =
+            (m['B'] as List<dynamic>).map((e) => Balance.fromMap(e)).toList();
 }
 
 class WsBalanceUpdate {
@@ -1461,5 +1540,7 @@ class WsListOrderStatus {
         listRejectReason = m['r'],
         listClientOrderId = m['C'],
         transactionTime = m['T'],
-        orders = (m['O'] as List<dynamic>).map((e) => WsOcoOrder.fromMap(e)).toList();
+        orders = (m['O'] as List<dynamic>)
+            .map((e) => WsOcoOrder.fromMap(e))
+            .toList();
 }