+++ lib/src/spot/endpoints/trade_endpoints.dart
@@ -60,7 +60,8 @@ extension TradeEndpoints on BinanceSpot {
     Map<String, String> params = {
       'symbol': symbol,
     };
-    if (origClientOrderId != null) params['origClientOrderId'] = origClientOrderId;
+    if (origClientOrderId != null)
+      params['origClientOrderId'] = origClientOrderId;
     if (orderId != null) params['orderId'] = orderId.toString().toString();
     if (newClientOrderId != null) params['newClientOrderId'] = newClientOrderId;
     if (recvWindow != null) params['recvWindow'] = recvWindow.toString();
@@ -71,7 +72,8 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(CanceledOrder.fromMap(r.right)));
+    ).then(
+        (r) => r.isLeft ? Left(r.left) : Right(CanceledOrder.fromMap(r.right)));
   }
 
   /// Cancels all active orders on a symbol.
@@ -91,7 +93,10 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<CanceledOrder>.from(r.right.map((e) => CanceledOrder.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<CanceledOrder>.from(
+            r.right.map((e) => CanceledOrder.fromMap(e)))));
   }
 
   /// Check an order's status.
@@ -105,7 +110,8 @@ extension TradeEndpoints on BinanceSpot {
       'symbol': symbol,
     };
     if (orderId != null) params['orderId'] = orderId.toString();
-    if (origClientOrderId != null) params['origClientOrderId'] = origClientOrderId;
+    if (origClientOrderId != null)
+      params['origClientOrderId'] = origClientOrderId;
     if (recvWindow != null) params['recvWindow'] = recvWindow.toString();
     return sendRequest(
       path: 'api/v3/order',
@@ -114,7 +120,8 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(CurrentOrder.fromMap(r.right)));
+    ).then(
+        (r) => r.isLeft ? Left(r.left) : Right(CurrentOrder.fromMap(r.right)));
   }
 
   /// Get all open orders on a symbol.
@@ -133,7 +140,10 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<CurrentOrder>.from(r.right.map((e) => CurrentOrder.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<CurrentOrder>.from(
+            r.right.map((e) => CurrentOrder.fromMap(e)))));
   }
 
   /// Get all account orders; active, canceled, or filled.
@@ -160,7 +170,10 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<CurrentOrder>.from(r.right.map((e) => CurrentOrder.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<CurrentOrder>.from(
+            r.right.map((e) => CurrentOrder.fromMap(e)))));
   }
 
   /// Send in a new OCO
@@ -187,13 +200,17 @@ extension TradeEndpoints on BinanceSpot {
       'price': price,
       'stopPrice': stopPrice,
     };
-    if (listClientOrderId != null) params['listClientOrderId'] = listClientOrderId;
-    if (limitClientOrderId != null) params['limitClientOrderId'] = limitClientOrderId.toString();
+    if (listClientOrderId != null)
+      params['listClientOrderId'] = listClientOrderId;
+    if (limitClientOrderId != null)
+      params['limitClientOrderId'] = limitClientOrderId.toString();
     if (limitIcebergQty != null) params['limitIcebergQty'] = limitIcebergQty;
-    if (stopClientOrderId != null) params['stopClientOrderId'] = stopClientOrderId;
+    if (stopClientOrderId != null)
+      params['stopClientOrderId'] = stopClientOrderId;
     if (stopLimitPrice != null) params['stopLimitPrice'] = stopLimitPrice;
     if (stopIcebergQty != null) params['stopIcebergQty'] = stopIcebergQty;
-    if (stopLimitTimeInForce != null) params['stopLimitTimeInForce'] = stopLimitTimeInForce.toStr();
+    if (stopLimitTimeInForce != null)
+      params['stopLimitTimeInForce'] = stopLimitTimeInForce.toStr();
     if (newOrderRespType != null) params['newOrderRespType'] = newOrderRespType;
     if (recvWindow != null) params['recvWindow'] = recvWindow.toString();
     return sendRequest(
@@ -218,7 +235,8 @@ extension TradeEndpoints on BinanceSpot {
       'symbol': symbol,
     };
     if (orderListId != null) params['orderListId'] = orderListId.toString();
-    if (listClientOrderId != null) params['listClientOrderId'] = listClientOrderId;
+    if (listClientOrderId != null)
+      params['listClientOrderId'] = listClientOrderId;
     if (newClientOrderId != null) params['newClientOrderId'] = newClientOrderId;
     if (recvWindow != null) params['recvWindow'] = recvWindow.toString();
     return sendRequest(
@@ -239,7 +257,8 @@ extension TradeEndpoints on BinanceSpot {
   }) {
     Map<String, String> params = {};
     if (orderListId != null) params['orderListId'] = orderListId.toString();
-    if (origClientOrderId != null) params['origClientOrderId'] = origClientOrderId;
+    if (origClientOrderId != null)
+      params['origClientOrderId'] = origClientOrderId;
     if (recvWindow != null) params['recvWindow'] = recvWindow.toString();
     return sendRequest(
       path: 'api/v3/orderList',
@@ -272,7 +291,9 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<OcoOrder>.from(r.right.map((e) => OcoOrder.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<OcoOrder>.from(r.right.map((e) => OcoOrder.fromMap(e)))));
   }
 
   /// Query all currently open OCO orders.
@@ -288,7 +309,9 @@ extension TradeEndpoints on BinanceSpot {
       signatureRequired: true,
       timestampRequired: true,
       params: params,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<OcoOrder>.from(r.right.map((e) => OcoOrder.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<OcoOrder>.from(r.right.map((e) => OcoOrder.fromMap(e)))));
   }
 
   /// Get current account information.
@@ -304,7 +327,8 @@ extension TradeEndpoints on BinanceSpot {
       signatureRequired: true,
       timestampRequired: true,
       params: params,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(AccountInfo.fromMap(r.right)));
+    ).then(
+        (r) => r.isLeft ? Left(r.left) : Right(AccountInfo.fromMap(r.right)));
   }
 
   /// Get trades for a specific account and symbol.
@@ -333,6 +357,9 @@ extension TradeEndpoints on BinanceSpot {
       keyRequired: true,
       signatureRequired: true,
       timestampRequired: true,
-    ).then((r) => r.isLeft ? Left(r.left) : Right(List<AccountTrade>.from(r.right.map((e) => AccountTrade.fromMap(e)))));
+    ).then((r) => r.isLeft
+        ? Left(r.left)
+        : Right(List<AccountTrade>.from(
+            r.right.map((e) => AccountTrade.fromMap(e)))));
   }
 }